code refactoring of session selection page
